# Code/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: '0.1.2'
  title: Simple tasks API
  description: An API example to test swagger
host: 192.168.99.100
basePath: /api
schemes:
  - http
paths:
  '/tasks':
    get:
      tags:
        - task
      summary: Get list of Tasks
      description: Presents the list of tasks existing in the system
      operationId: getTasks
      produces:
        - application/json
      responses:
        '200':
          description: Information returned correctly
          schema:
            type: array
            items:
              $ref: '#/definitions/Task'
    post:
      tags:
        - task
      summary: Creation of a new task
      description: Creation of a new task
      operationId: postTask
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "The task details"
          required: true
          schema:
            $ref: '#/definitions/NewTask'
      responses:
        '201':
          description: "Information returned correctly"
          schema:
            $ref: '#/definitions/Task'
        '415':
          description: "Payload sent does not respect accept header"
          schema:
            $ref: '#/definitions/ErrorResponse'
        '422':
          description: "Json is incomplete or doesn't match the application's specification"
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/tasks/{task_id}':
    get:
      tags:
        - task
      summary: Get a task
      description: Get the task with the corresponding id
      operationId: getTaskById
      parameters:
      - in: "path"
        name: "task_id"
        description: "Task ID"
        required: true
        type: "string"
      produces:
        - application/json
      responses:
        '200':
          description: "Information returned correctly"
          schema:
            $ref: '#/definitions/Task'
        '404':
          description: "Task wanted was not found"
  '/executions':
    post:
      tags:
        - execution
      summary: Creation of a new execution
      description: Creation of a new execution
      operationId: postExecution
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "The execution details"
          required: true
          schema:
            $ref: '#/definitions/NewExecution'
      responses:
        '201':
          description: "Information returned correctly"
          schema:
            $ref: '#/definitions/Execution'
        '415':
          description: "Payload sent does not respect accept header"
          schema:
            $ref: '#/definitions/ErrorResponse'
        '422':
          description: "Json is incomplete or doesn't match the application's specification"
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/executions/{execution_id}':
    get:
      tags:
        - execution
      summary: Get an execution
      description: Get the execution with the corresponding id
      operationId: getExecutionById
      parameters:
        - in: "path"
          name: execution_id
          description: Task ID
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Information returned correctly
          schema:
            $ref: '#/definitions/Execution'
        '404':
          description: "Step wanted was not found"
  '/steps':
    post:
      tags:
        - step
      summary: Creation of a new step
      description: Creation of a new step
      operationId: postStep
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "The step details"
          required: true
          schema:
            $ref: '#/definitions/NewStep'
      responses:
        '201':
          description: "Information returned correctly"
          schema:
            $ref: '#/definitions/Step'
        '415':
          description: "Payload sent does not respect accept header"
          schema:
            $ref: '#/definitions/ErrorResponse'
        '422':
          description: "Json is incomplete or doesn't match the application's specification"
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/steps/{step_id}':
    get:
      tags:
        - step
      summary: Get an execution
      description: Get the execution with the corresponding id
      operationId: getStepById
      parameters:
        - in: "path"
          name: step_id
          description: Task ID
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: "Information returned correctly"
          schema:
            $ref: '#/definitions/Step'
        '404':
          description: "Step wanted was not found"
definitions:
  ErrorResponse:
    type: object
    properties:
      success:
        type: boolean
      error:
        type: string
  NewTask:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      task_id:
        type: string
    required: [name]
  Task:
    type: object
    properties:
      id:
        type: string
      task_id:
        type: string
      self:
        type: string
      name:
        type: string
      description:
        type: string
      creationDate:
        type: integer
        format: int64
      executions:
        type: array
        items:
          $ref: '#/definitions/Execution'
  NewExecution:
    type: object
    properties:
      task_id:
        type: string
      name:
        type: string
    required: [task_id]
  Execution:
    type: object
    properties:
      id:
        type: string
      self:
        type: string
      name:
        type: string
      creationDate:
        type: integer
        format: int64
      steps:
        type: array
        items:
          $ref: '#/definitions/Step'
  NewStep:
    type: object
    properties:
      execution_id:
        type: string
      name:
        type: string
      context:
        type: string
      status:
        type: string
        enum:
          - Running
          - Error
          - Log
          - Success
    required: [execution_id, name, status]
  Step:
    type: object
    properties:
      id:
        type: string
      creationDate:
        type: integer
        format: int64
      name:
        type: string
      context:
        type: string
      status:
        type: string